QA: skip test when minimum PHP version does not comply
The `DateTimeImmutable` class was only introduced in PHP 5.5.0.
The `DateInterval` class in PHP 5.3.0.

This test can therefore not be run on PHP < 5.5.0.

Adding the PHPUnit `@requires` annotation will ensure that the test suite will not break on this test.

Alternatively, if testing this functionality on PHP 5.2 - 5.4 is desired, the test will need to be refactored and the above mentioned classes removed from the test.

Ref:
* https://phpunit.de/manual/4.8/en/incomplete-and-skipped-tests.html#incomplete-and-skipped-tests.requires.tables.api